+++ mayan/settings/development/base.py
@@ -11,16 +11,6 @@ CELERY_TASK_ALWAYS_EAGER = True
 CELERY_TASK_EAGER_PROPAGATES = CELERY_TASK_ALWAYS_EAGER  # NOQA: F405
 EMAIL_BACKEND = 'django.core.mail.backends.console.EmailBackend'
 
-if 'rosetta' not in INSTALLED_APPS:   # NOQA: F405
-    try:
-        import rosetta  # NOQA: F401
-    except ImportError:
-        pass
-    else:
-        INSTALLED_APPS += (  # NOQA: F405
-            'rosetta',
-        )
-
 if 'django_extensions' not in INSTALLED_APPS:
     try:
         import django_extensions  # NOQA: F401